Get rid of "fake warnings".
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53467 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/ncbi_conn_streambuf.cpp
@@ -262,7 +262,7 @@ streamsize CConn_Streambuf::xsputn(const CT_CHAR_TYPE* buf, streamsize m)
 
     if (m <= 0)
         return 0;
-    _ASSERT(m < numeric_limits<size_t>::max());  // NCBI_FAKE_WARNING
+    _ASSERT((Uint8)m < numeric_limits<size_t>::max());
     size_t n = (size_t) m;
 
     m_Status = eIO_Success;
@@ -385,7 +385,7 @@ streamsize CConn_Streambuf::xsgetn(CT_CHAR_TYPE* buf, streamsize m)
 
     if (m <= 0)
         return 0;
-    _ASSERT(m < numeric_limits<size_t>::max());  // NCBI_FAKE_WARNING
+    _ASSERT((Uint8)m < numeric_limits<size_t>::max());
     size_t n = (size_t) m;
 
     // first, read from the memory buffer
