+++ src/corelib/rwstreambuf.cpp
@@ -284,7 +284,7 @@ streamsize CRWStreambuf::xsputn(const CT_CHAR_TYPE* buf, streamsize m)
 
     if (m <= 0)
         return 0;
-    _ASSERT(m < numeric_limits<size_t>::max());  // NCBI_FAKE_WARNING
+    _ASSERT((Uint8)m < numeric_limits<size_t>::max());
     size_t n = (size_t) m;
 
     ERW_Result result = eRW_Success;
@@ -407,7 +407,7 @@ streamsize CRWStreambuf::xsgetn(CT_CHAR_TYPE* buf, streamsize m)
 
     if (m <= 0)
         return 0;
-    _ASSERT(m < numeric_limits<size_t>::max());  // NCBI_FAKE_WARNING
+    _ASSERT((Uint8)m < numeric_limits<size_t>::max());
     size_t n = (size_t) m;
 
     // first, read from the memory buffer
